Update internetIncome.sh to add spide.network
Added spide.network service.
You first have to start the node and it generates two keys, a machine key and a device key. You have to paste the device key on the website to register the node. The machine key has to be used to start the container again with the same machine and device key.
I am saving them into two separate files that are backed up. spide_machine.txt is necessary to restart the node(s) and spide_device.txt is handy to have all IDs in one file to register them on their website.